Fixed parsing of duration for EXT-X-CUE-OUT (#39)

+++ src/main/java/io/lindstrom/m3u8/model/MediaSegment.java
@@ -31,7 +31,7 @@ public interface MediaSegment {
         return false;
     }
 
-    Optional<Integer> cueOut();
+    Optional<Double> cueOut();
 
     @Value.Default
     default boolean cueIn() {
